backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3.1,<6.0
djongo>=1.3.6,<1.3.8
sqlparse==0.2.4
pymongo>=4.6,<5.0
Pillow>=10.0,<11.0
numpy>=1.26,<2.1
opencv-python>=4.9,<5.0
tensorflow
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
django-cors-headers>=4.3,<5.0
drf-spectacular
torch
torchvision
# Pour CUDA, ajoute torch==<version>+cu118 selon ta config, sinon CPU
# Ex: torch==2.2.0+cu118 torchvision==0.17.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html

# Detectron2 depuis le repo officiel (CPU)
git+https://github.com/facebookresearch/detectron2.git

# Add any other package you use, with similar version pinning